完成管理员服务

+++ app/administrator/service/internal/biz/administrator.go
@@ -3,38 +3,39 @@ package biz
 import (
 	"context"
 	v1 "github.com/ZQCard/kratos-base-project/api/administrator/v1"
-	"github.com/ZQCard/kratos-base-project/app/administrator/service/internal/data/entity"
+	"github.com/ZQCard/kratos-base-project/pkg/errResponse"
+	"net/http"
+
+	"github.com/ZQCard/kratos-base-project/pkg/utils/typeConvert"
+	"github.com/ZQCard/kratos-base-project/pkg/validate"
+
 	"github.com/go-kratos/kratos/v2/errors"
 	"github.com/go-kratos/kratos/v2/log"
 )
 
-var (
-	ErrRecordNotFound       = "管理员不存在"
-	ErrAdministratorForbid  = "管理员已禁用"
-	ErrAdministratorDeleted = "管理员已删除"
-)
-
 type Administrator struct {
 	Id        int64
-	Username  string
-	Password  string
-	Mobile    string
-	Nickname  string
-	Avatar    string
-	Status    int64
-	IsDeleted bool
+	Username  string `validate:"required,max=50" label:"用户名"`
+	Password  string `validate:"required,max=50" label:"密码"`
+	Salt      string
+	Mobile    string `validate:"required,max=20" label:"手机号码"`
+	Nickname  string `validate:"required,max=50" label:"昵称"`
+	Avatar    string `validate:"required,max=150" label:"头像地址"`
+	Status    int64  `validate:"required,oneof=1 2" label:"状态"`
 	CreatedAt string
 	UpdatedAt string
 	DeletedAt string
 }
 
+// AdministratorRepo 模块接口
 type AdministratorRepo interface {
-	VerifyPassword(ctx context.Context, id int64, password string) (bool, error)
 	CreateAdministrator(ctx context.Context, reqData *Administrator) (*Administrator, error)
 	UpdateAdministrator(ctx context.Context, reqData *Administrator) (*Administrator, error)
 	GetAdministrator(ctx context.Context, params map[string]interface{}) (*Administrator, error)
-	ListAdministrator(ctx context.Context, pageNum, pageSize int64) ([]*Administrator, int64, error)
+	ListAdministrator(ctx context.Context, pageNum, pageSize int64, params map[string]interface{}) ([]*Administrator, int64, error)
 	DeleteAdministrator(ctx context.Context, id int64) error
+	RecoverAdministrator(ctx context.Context, id int64) error
+	VerifyAdministratorPassword(ctx context.Context, id int64, password string) (bool, error)
 }
 
 type AdministratorUseCase struct {
@@ -43,57 +44,53 @@ type AdministratorUseCase struct {
 }
 
 func NewAdministratorUseCase(repo AdministratorRepo, logger log.Logger) *AdministratorUseCase {
-	logs := log.NewHelper(log.With(logger, "module", "administrator/interface"))
-	return &AdministratorUseCase{
-		repo: repo,
-		log:  logs,
-	}
-}
-
-func (ac AdministratorUseCase) FindLoginAdministratorByUsername(ctx context.Context, in *v1.GetLoginAdministratorByUsernameRequest) (*v1.GetLoginAdministratorByUsernameReply, error) {
-	params := make(map[string]interface{})
-	params["username"] = in.Username
-	administrator, err := ac.repo.GetAdministrator(ctx, params)
-	if err != nil {
-		return &v1.GetLoginAdministratorByUsernameReply{}, err
-	}
-	if administrator.Status == entity.AdministratorStatusForbid {
-		return &v1.GetLoginAdministratorByUsernameReply{}, errors.New(400, "ADMINISTRATOR_FORBIDDEN", ErrAdministratorForbid)
-	}
-	if administrator.DeletedAt != "" {
-		return &v1.GetLoginAdministratorByUsernameReply{}, errors.New(400, "ADMINISTRATOR_DELETED", ErrAdministratorDeleted)
-	}
-
-	return &v1.GetLoginAdministratorByUsernameReply{
-		Id:       administrator.Id,
-		Username: administrator.Username,
-	}, nil
-}
-
-func (ac AdministratorUseCase) VerifyAdministratorPassword(ctx context.Context, in *v1.VerifyPasswordRequest) (bool, error) {
-	result, err := ac.repo.VerifyPassword(ctx, in.Id, in.Password)
-	if err != nil {
-		return false, err
-	}
-	return result, nil
+	return &AdministratorUseCase{repo: repo, log: log.NewHelper(log.With(logger, "module", "usecase/beer"))}
 }
 
 func (uc *AdministratorUseCase) Create(ctx context.Context, data *Administrator) (*Administrator, error) {
+	err := validate.ValidateStructCN(data)
+	if err != nil {
+		return &Administrator{}, errors.New(http.StatusBadRequest, errResponse.ReasonParamsError, err.Error())
+	}
 	return uc.repo.CreateAdministrator(ctx, data)
 }
 
 func (uc *AdministratorUseCase) Delete(ctx context.Context, id int64) error {
+	if id == 0 {
+		return errResponse.SetCustomizeErrInfoByReason(errResponse.ReasonMissingId)
+	}
 	return uc.repo.DeleteAdministrator(ctx, id)
 }
 
+func (uc *AdministratorUseCase) Recover(ctx context.Context, id int64) error {
+	if id == 0 {
+		return errResponse.SetCustomizeErrInfoByReason(errResponse.ReasonMissingId)
+	}
+	return uc.repo.RecoverAdministrator(ctx, id)
+}
+
 func (uc *AdministratorUseCase) Update(ctx context.Context, data *Administrator) (*Administrator, error) {
+	if data.Id == 0 {
+		return &Administrator{}, errResponse.SetCustomizeErrInfoByReason(errResponse.ReasonMissingId)
+	}
 	return uc.repo.UpdateAdministrator(ctx, data)
 }
 
 func (uc *AdministratorUseCase) Get(ctx context.Context, params map[string]interface{}) (*Administrator, error) {
+	params = typeConvert.ClearMapZeroValue(params)
 	return uc.repo.GetAdministrator(ctx, params)
 }
 
-func (uc *AdministratorUseCase) List(ctx context.Context, pageNum, pageSize int64) ([]*Administrator, int64, error) {
-	return uc.repo.ListAdministrator(ctx, pageNum, pageSize)
+func (uc *AdministratorUseCase) List(ctx context.Context, pageNum, pageSize int64, params map[string]interface{}) ([]*Administrator, int64, error) {
+	params = typeConvert.ClearMapZeroValue(params)
+	return uc.repo.ListAdministrator(ctx, pageNum, pageSize, params)
+}
+
+func (ac AdministratorUseCase) VerifyAdministratorPassword(ctx context.Context, in *v1.VerifyAdministratorPasswordRequest) (bool, error) {
+
+	result, err := ac.repo.VerifyAdministratorPassword(ctx, in.Id, in.Password)
+	if err != nil {
+		return false, err
+	}
+	return result, nil
 }
